requirements.txt
# Tested on python version 3.9.23


# Scientific and data manipulation
numpy==1.21.6
pytorch-lightning==1.5.10
scipy
pandas
h5py
scikit-learn
trimesh==3.23.3

scikit-image
scikit-video

# Visualization and 3D
matplotlib
plyfile
pyglet==1.5.27
open3d
rtree
pyrender
seaborn


# Utilities
tqdm
jupyter
imageio
imageio-ffmpeg
einops
omegaconf
natsort
hydra-core
torchmetrics==0.9.3
torch-fidelity
setuptools<81

# Training Visualization
wandb==0.15.8
tensorboard